fix: Fix the Size dataType issue

See merge request urgent-couriers/runbuilder!6

+++ Models/UTL_stpJob_tblBulkJobResult.cs
@@ -39,7 +39,7 @@ namespace RunBuilder.Models
         [Column("Amount", TypeName = "money")]
         public decimal Amount { get; set; }
         public int SpeedID { get; set; }
-        public short? Size { get; set; }
+        public int? Size { get; set; }
         public short? Items { get; set; }
         [Column("Weight", TypeName = "decimal(6,2)")]
         public decimal? Weight { get; set; }
